Place

+++ backendHackathon/model/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, DateTime, ForeignKey
+from sqlalchemy import Column, String, Integer, DateTime, ForeignKey, Float
 from model import Base
 from datetime import datetime
 
@@ -9,7 +9,7 @@ class User(Base):
     password = Column(String(64), nullable=False)  # hashedpass
     name = Column(String(100), nullable=False)
     email = Column(String(100), nullable=False)
-    localizacao_lat = Column(Integer(), nullable=True)
-    localizacao_lon = Column(Integer(), nullable=True)
+    localizacao_lat = Column(Float(), nullable=True)
+    localizacao_lon = Column(Float(), nullable=True)
     created_on = Column(DateTime(), default=datetime.now, nullable=False)
     team = Column(Integer(), ForeignKey('team.id'), nullable=True)
